folders from other users can be added to own dashboard and removed. users that do not own the folder cannot change the layout, privacy,  or edit/remove its elements or the folder itself. in dashboard folders from other users are displayed with a folder *by username*

+++ back/routes/dashboard.js
@@ -11,7 +11,9 @@ const { removeCloudinaryFolder } = require("../middleware/cloudinary")
 //RETRIEVE FOLDERS FROM USER
 router.get("/folders", async (req, res, next) => {
   if (req.user) {
-    const { folders } = await User.findOne({ _id: req.user._id }).populate("folders");
+    const { folders } = await User.findOne({ _id: req.user._id }).populate({
+      path: 'folders', populate: { path: 'user', select: { '_id': 1, 'username': 1 } }
+    })
     res.status(200).json(folders)
   } else {
     res.status(401)
@@ -79,7 +81,6 @@ router.post("/create/:folder", async (req, res, next) => {
   if (req.user) {
     const { folder } = req.params;
     try {
-
       // No duplicated folder names
       const existingFolder = await Folder.findOne({ $and: [{ user: req.user._id }, { folder }] })
       if (existingFolder !== null) {
@@ -96,8 +97,29 @@ router.post("/create/:folder", async (req, res, next) => {
   } else {
     res.status(401)
   }
-
 })
+
+//ADD FOLDER FROM ANOTHER USER TO OWN DASHBOARD
+router.post(`/add/folder/:username/:folder`, async (req, res, next) => {
+  if (req.user) {
+    const { folder, username } = req.params;
+    try {
+      await Folder.findOne({ path: `/${username}/${folder.replace(/ /g, "_")}` }, async (err, res) => {
+
+        if (res.isPrivate == false) { User.updateOne({ _id: req.user._id }, { $push: { folders: res._id } }, async (err, res) => console.log(res)) }
+      });
+      res.status(200).json({ message: `${folder} folder created` });
+    }
+    catch {
+      res.status(500).json({ message: "Something went wrong" })
+    }
+  } else {
+    res.status(401)
+  }
+})
+
+
+
 
 // CHANGE FOLDER PRIVACY 
 router.post(`/:folder/privacy`, async (req, res, next) => {
@@ -117,19 +139,26 @@ router.post(`/:folder/privacy`, async (req, res, next) => {
 })
 
 // DELETE FOLDER and its elements FROM DB
-router.delete("/:folder", async (req, res, next) => {
+// OR REMOVE FOLDER FROM DASHBOARD IF USER IS NOT OWNER
+router.post("/delete/folder", async (req, res, next) => {
   if (req.user) {
-    const { folder } = req.params;
+    const { folder } = req.body;
+    console.log("backroute", folder)
     try {
-      await Folder.findOneAndDelete({
-        $and: [{ user: req.user._id }, { path: `/${req.user.username}/${folder.replace(/ /g, "_")}` }]
-      }, async (err, res) => {
-        await User.updateOne({ _id: req.user._id }, { $pull: { folders: res._id } })
-        await Element.deleteMany({ $and: [{ user: req.user._id }, { folder: res._id }] });
-        //delete also pics from cloudinary
-        removeCloudinaryFolder({ id: req.user._id, folder });
+      if (folder.user._id == req.user._id) {
+        await Folder.findOneAndDelete({
+          $and: [{ user: req.user._id }, { path: `/${req.user.username}/${folder.replace(/ /g, "_")}` }]
+        }, async (err, res) => {
+          await User.updateOne({ _id: req.user._id }, { $pull: { folders: res._id } })
+          await Element.deleteMany({ $and: [{ user: req.user._id }, { folder: res._id }] });
+          //delete also pics from cloudinary
+          removeCloudinaryFolder({ id: req.user._id, folder });
 
-      });
+        });
+      }
+      else {
+        await User.updateOne({ _id: req.user._id }, { $pull: { folders: folder._id } })
+      }
       res.status(200).json({ message: `${folder} folder deleted` });
     }
     catch (err) {
@@ -158,11 +187,11 @@ router.post("/update/:folder/layout", async (req, res, next) => {
 });
 
 //GET FOLDER LAYOUT
-router.get("/:folder/layout", async (req, res, next) => {
-  const { folder } = req.params;
+router.get("/:username/:folder/layout", async (req, res, next) => {
+  const { folder, username } = req.params;
   if (req.user) {
     try {
-      const { layout } = await Folder.findOne({ $and: [{ user: req.user._id }, { path: `/${req.user.username}/${folder.replace(/ /g, "_")}` }] });
+      const { layout } = await Folder.findOne({ path: `/${username}/${folder.replace(/ /g, "_")}` });
       res.status(200).json(layout);
     }
     catch {
